+++ src/models/partner_user_position.py
@@ -8,17 +8,7 @@ from uuid import UUID, uuid4
 import sqlalchemy as sa
 from sqlmodel import Field, SQLModel
 
-class ProtocolType(str, Enum):
-    """ High-level category of the protocol for application logic. """
-    DEX_UNISWAPV3 = "DEX_UNISWAPV3"
-    LENDING_HYPURRFI = "LENDING_HYPURRFI"
-    YIELD_PENDLE = "YIELD_PENDLE"
-
-class QuantityType(str, Enum):
-    """ The specific type of financial activity that earns points. """
-    LP = "LP"
-    YT = "YT"
-    BORROW = "BORROW"
+from .enums import ProtocolType, QuantityType
 
 class PartnerUserPosition(SQLModel, table=True):
     """
